+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - resultBroad dataProfile.
 - added organization summary.
 - added period of record data profile with CSV output.
+- added Biological Metric data profile
 
 ### Changed
 - Station summary service returns only sites that have summary information.